Add PostPage component
In preparation for the use of HOC for conditional rendering, PostPage component was created to take the heavy lifting currently being done by the App component. This is to make sure that the only responsibility given to the App component is conditional rendering.
