+++ instagram-app/src/App.css
@@ -77,7 +77,3 @@
   font-size: 1.2rem;
   outline: none;
 }
-
-/* .comment-container  {
-  border-bottom: 1px solid #ccc;
-} */
